[OTA] Move OTARequestorDriver to clusters directory
