+++ examples/ota-requestor-app/esp32/main/main.cpp
@@ -30,6 +30,7 @@
 #include "nvs_flash.h"
 #include <app/clusters/network-commissioning/network-commissioning.h>
 #include <app/clusters/ota-requestor/BDXDownloader.h>
+#include <app/clusters/ota-requestor/GenericOTARequestorDriver.h>
 #include <app/clusters/ota-requestor/OTARequestor.h>
 #include <app/server/Server.h>
 #include <platform/ESP32/NetworkCommissioningDriver.h>
@@ -40,7 +41,6 @@
 #include <lib/support/ErrorStr.h>
 
 #include "OTAImageProcessorImpl.h"
-#include "platform/GenericOTARequestorDriver.h"
 
 using namespace ::chip;
 using namespace ::chip::System;
