+++ examples/platform/mbed/ota/OTARequestorDriverImpl.h
@@ -22,8 +22,8 @@
 
 #pragma once
 
-#include <GenericOTARequestorDriver.h>
-#include <OTARequestorDriver.h>
+#include <app/clusters/ota-requestor/GenericOTARequestorDriver.h>
+#include <app/clusters/ota-requestor/OTARequestorDriver.h>
 #include <app/clusters/ota-requestor/OTARequestorInterface.h>
 #include <lib/core/CHIPCallback.h>
 #include <lib/core/CHIPError.h>
