+++ examples/platform/qpg/ota/ota.cpp
@@ -28,8 +28,8 @@
 #include <platform/CHIPDeviceLayer.h>
 
 #include <app/clusters/ota-requestor/BDXDownloader.h>
+#include <app/clusters/ota-requestor/GenericOTARequestorDriver.h>
 #include <app/clusters/ota-requestor/OTARequestor.h>
-#include <platform/GenericOTARequestorDriver.h>
 #include <platform/qpg/OTAImageProcessorImpl.h>
 
 using namespace chip;
