+++ src/app/chip_data_model.gni
@@ -132,6 +132,8 @@ template("chip_data_model") {
           "${_app_root}/clusters/${cluster}/BDXDownloader.cpp",
           "${_app_root}/clusters/${cluster}/BDXDownloader.h",
           "${_app_root}/clusters/${cluster}/DefaultOTARequestorUserConsentProvider.h",
+          "${_app_root}/clusters/${cluster}/ExtendedOTARequestorDriver.cpp",
+          "${_app_root}/clusters/${cluster}/GenericOTARequestorDriver.cpp",
           "${_app_root}/clusters/${cluster}/OTARequestor.cpp",
         ]
       } else if (cluster == "bindings") {
