+++ src/app/clusters/ota-requestor/GenericOTARequestorDriver.cpp
@@ -29,11 +29,11 @@
 // - The provider location passed in AnnounceOTAProviders is used in a single query (possibly retried) and then discarded
 // - Explicitly triggering a query through TriggerImmediateQuery() cancels any in-progress update
 
-#include "GenericOTARequestorDriver.h"
-
-#include <app/clusters/ota-requestor/OTARequestorInterface.h>
 #include <platform/CHIPDeviceLayer.h>
 #include <platform/OTAImageProcessor.h>
+
+#include "GenericOTARequestorDriver.h"
+#include "OTARequestorInterface.h"
 
 namespace chip {
 namespace DeviceLayer {
